Fix build issues for onnxruntime_providers_cuda.
