+++ onnxruntime/contrib_ops/cuda/inverse.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/providers/cuda/cuda_common.h"
+#include "core/providers/cuda/cuda_kernel.h"
 #include "core/providers/cuda/math/unary_elementwise_ops_impl.h"
 
 namespace onnxruntime {
