+++ onnxruntime/core/providers/cuda/cuda_common.h
@@ -4,7 +4,6 @@
 #pragma once
 
 #include "core/common/status.h"
-#include "core/providers/cuda/cuda_fwd.h"
 #include "core/providers/cuda/cuda_pch.h"
 #include "core/providers/cuda/shared_inc/cuda_call.h"
 #include "core/providers/cuda/shared_inc/fast_divmod.h"
