+++ onnxruntime/core/providers/cuda/cuda_fence.cc
@@ -1,9 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "cuda_common.h"
-#include "cuda_fence.h"
-#include "gpu_data_transfer.h"
+#include "core/providers/cuda/cuda_fence.h"
+
+#include "core/graph/constants.h"
+#include "core/providers/cuda/cuda_common.h"
+#include "core/providers/cuda/gpu_data_transfer.h"
 
 namespace onnxruntime {
