+++ onnxruntime/core/providers/cuda/cuda_kernel.h
@@ -6,7 +6,9 @@
 #include "core/common/status.h"
 #include "core/framework/data_transfer_manager.h"
 #include "core/framework/op_kernel.h"
+#include "core/providers/cuda/cuda_common.h"
 #include "core/providers/cuda/cuda_execution_provider.h"
+#include "core/providers/cuda/cuda_fwd.h"
 
 namespace onnxruntime {
 namespace cuda {
