+++ onnxruntime/core/providers/cuda/nn/batch_norm.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "gsl/gsl"
+#include "core/providers/cuda/cuda_kernel.h"
 #include "core/providers/cuda/cudnn_common.h"
 
 namespace onnxruntime {
