+++ onnxruntime/core/providers/cuda/nn/conv.h
@@ -5,7 +5,7 @@
 
 #include "core/common/common.h"
 #include "core/platform/ort_mutex.h"
-#include "core/framework/op_kernel.h"
+#include "core/providers/cuda/cuda_kernel.h"
 #include "core/providers/cuda/cudnn_common.h"
 #include "core/providers/cpu/nn/conv_attributes.h"
 #include <list>
