+++ onnxruntime/core/providers/cuda/nn/lrn.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "core/framework/op_kernel.h"
+#include "core/providers/cuda/cuda_kernel.h"
 #include "core/providers/cuda/cudnn_common.h"
 
 namespace onnxruntime {
