+++ onnxruntime/core/providers/cuda/rnn/cudnn_rnn_base.h
@@ -4,9 +4,11 @@
 #pragma once
 
 #include "gsl/gsl"
-#include "core/providers/cuda/cudnn_common.h"
-#include "core/providers/cuda/cuda_kernel.h"
+
 #include <cudnn.h>
+
+#include "core/providers/cuda/cuda_kernel.h"
+#include "core/providers/cuda/cudnn_common.h"
 
 namespace onnxruntime {
 namespace cuda {
