+++ orttraining/orttraining/training_ops/cuda/nn/batch_norm_grad.h
@@ -4,6 +4,8 @@
 #pragma once
 
 #include "gsl/gsl"
+
+#include "core/providers/cuda/cuda_kernel.h"
 #include "core/providers/cuda/cudnn_common.h"
 
 namespace onnxruntime {
